+++ src/borrow_pcg/borrow_pcg_edge.rs
@@ -12,7 +12,7 @@ use super::{
     has_pcs_elem::HasPcsElems,
     path_condition::{PathCondition, PathConditions},
     region_projection::{MaybeRemoteRegionProjectionBase, RegionProjection},
-    edge::region_projection_member::BlockEdge,
+    edge::block::BlockEdge,
 };
 use crate::borrow_pcg::edge::abstraction::AbstractionType;
 use crate::borrow_pcg::edge::borrow::BorrowEdge;
