+++ src/borrow_pcg/edge/mod.rs
@@ -1,4 +1,4 @@
 pub mod borrow;
 pub mod abstraction;
 pub mod kind;
-pub mod region_projection_member;
+pub mod block;
